fix(docs): split block page into server/client components
Separate generateStaticParams (server) from React.lazy (client)
to fix "use client" + generateStaticParams conflict.
